refactor(data): extract shared GHS utilities into ghs-utils.ts
Move label resolution, game token resolution, action formatting, and
common types from import-character-abilities.ts into a shared module.
Move kebabToTitle from import-monster-stats.ts. Add loadLabels() helper
and GHS_DATA_DIR constant so future import scripts can reuse them
without duplicating label-loading logic.
